bug fix: add empty unit check for adding overtime; detailed label on entry form; full day name

+++ utils/formatDate.js
@@ -20,5 +20,5 @@ export function getDayName(dateStr, format = "YYYY-MM-DD") {
     );
   }
 
-  return date.toLocaleDateString("en-US", { weekday: "short" }).toUpperCase();
+  return date.toLocaleDateString("en-US", { weekday: "long" }).toUpperCase();
 }
